backend/src/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.6.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.1
dnspython==2.2.1
email-validator==1.3.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
evmdasm==0.1.10
fastapi==0.88.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
grevillea==0.0.1
h11==0.14.0
hexbytes==0.3.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.8.1
peewee==3.15.4
protobuf==3.19.5
psycopg2==2.9.5
pycryptodome==3.16.0
pydantic==1.10.2
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toolz==0.12.0
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.3
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
peewee-migrations

# Test dependencies
pytest
freezegun
